Use Type=notify for daemon service
Type="oneshot" in combination with continuing to run the daemon in the same
thread post-startup makes the systemd service get stuck on activating.

This is bad for getting nvidia-exec to work on NixOS, because `nixos-rebuild`
demands that all services are activated successfully.

This uses the sd_notify protocol to notify systemd once the daemon is running.
The daemon will not fail if NOTIFY_SOCKET is unavailable, keeping
compatibility with systems without systemd.